ASTS/src/anomaly_avoidance.py
# Filename: anomaly_avoidance.py

import random

def enhanced_anomaly_avoidance():
    """
    Implement enhanced anomaly avoidance techniques.
    This function can include various strategies to make the bot behavior less predictable.
    """
    print("Enhanced anomaly avoidance active")

def random_jitter(position, amount=5):
    """
    Add random jitter to the position to avoid detection.
    
    Args:
    position (tuple): The current (x, y) position.
    amount (int): The maximum jitter amount. Default is 5.

    Returns:
    tuple: The jittered position.
    """
    jittered_position = (position[0] + random.uniform(-amount, amount),
                         position[1] + random.uniform(-amount, amount))
    return jittered_position

def occasional_mistake(success_rate=0.95):
    """
    Occasionally make a mistake to mimic human behavior.
    
    Args:
    success_rate (float): The probability of not making a mistake. Default is 0.95.

    Returns:
    bool: True if no mistake is made, False otherwise.
    """
    return random.random() < success_rate