ASTS/src/tests/test_anomaly_avoidance.py
import unittest
from src.anomaly_avoidance import enhanced_anomaly_avoidance, random_jitter, occasional_mistake

class TestAnomalyAvoidance(unittest.TestCase):

    def test_enhanced_anomaly_avoidance(self):
        # Assuming the function should return True if anomaly is avoided
        self.assertTrue(enhanced_anomaly_avoidance())

    def test_random_jitter(self):
        # Test if random jitter returns a value within an expected range
        jitter_value = random_jitter(10)
        self.assertGreaterEqual(jitter_value, -10)
        self.assertLessEqual(jitter_value, 10)

    def test_occasional_mistake(self):
        # Test if occasional mistake is making the expected decision
        mistake_value = occasional_mistake(0.1)  # 10% chance to make a mistake
        self.assertIn(mistake_value, [True, False])  # Should return a boolean value

if __name__ == '__main__':
    unittest.main()
